requirements.txt
Django==3.0.4
psycopg2-binary==2.8.4
pylint==2.3.1
pylint-django==2.0.8
requests==2.23.0
graphene==2.1.8
graphene-django==2.8.2
django-graphql-jwt==0.3.0
